+++ src/main/java/dk/gokartland/booking/domain/RestaurantBooking.java
@@ -7,7 +7,7 @@ import java.util.Date;
  */
 public class RestaurantBooking extends FacilityBooking {
 
-        public RestaurantBooking(Date from, Date to, String comments, short numberOfPeople, BookablePlace bookablePlace) {
+        public RestaurantBooking(Date from, Date to, String comments, int numberOfPeople, BookablePlace bookablePlace) {
         super(from, to, comments, numberOfPeople, bookablePlace);
     }
 }
